requirements.txt
boto3==1.24.75
pandas~=1.4.4
awswrangler~=2.16.1
numpy~=1.23.2
requests~=2.28.0
pyfunctional~=1.4.3